Revert "feat:ファイル整理とTitleSceneに機能追加"
This reverts commit de6bc77316653e45c683bf27c3a7dc2d3bf02937.
