--- Assets/Bgm.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9c4f8fa6cf700d2428cb7796f3dfcab6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
