--- Assets/Material.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c9ff04dbc11a056418d456f2e2b6b727
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
