--- Assets/Prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0aa36b3f54223d640b91a9b507dffe07
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
