+++ Assets/Scenes/ClearScene.cs
