--- Assets/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff46994e098f1da44ada0ff54a7876fb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
