--- Assets/Scripts/AudioSetting.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 40f19679b4b8f8b458abf984169678e5
